Add files via upload
About this file

Add Suggestion
This dataset contains insights into credit card transactions made in India, offering a comprehensive look at the spending habits of Indians across the nation

City: The city in which the transaction took place. (String)
Date: The date of the transaction. (Date)
Card Type: The type of credit card used for the transaction. (String)
Exp Type: The type of expense associated with the transaction. (String)
Gender: The gender of the cardholder. (String)
Amount: The amount of the transaction. (Number)